perception/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
google-cloud-vision==3.7.2
scikit-learn==1.5.1
requests==2.32.3
pydantic==2.8.2
